Big refactor - part 2
