+++ FeatureAuthorize/CalcFeaturePermissions.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) 2019 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
-using System;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
@@ -9,7 +8,7 @@ using DataLayer.EfCode;
 using Microsoft.EntityFrameworkCore;
 using PermissionParts;
 
-namespace ServiceLayer.CodeCalledInStartup
+namespace FeatureAuthorize
 {
     /// <summary>
     /// This is the code that calculates what feature permissions a user has
