+++ PermissionAccessControl2/Startup.cs
@@ -20,7 +20,6 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using ServiceLayer.AppStart;
 using ServiceLayer.AuthoriseSetup;
-using ServiceLayer.CodeCalledInStartup;
 using ServiceLayer.UserServices;
 using Swashbuckle.AspNetCore.Swagger;
