+++ ServiceLayer/AuthoriseSetup/AddPermissionsToUserClaims.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) 2019 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
-using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using DataAuthorize;
@@ -11,7 +10,7 @@ using FeatureAuthorize;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Options;
 
-namespace ServiceLayer.CodeCalledInStartup
+namespace ServiceLayer.AuthoriseSetup
 {
     /// <summary>
     /// This version provides:
