+++ ServiceLayer/AuthoriseSetup/AuthCookieValidatePermissionsOnly.cs
@@ -9,7 +9,6 @@ using DataLayer.EfCode;
 using FeatureAuthorize;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.Extensions.DependencyInjection;
-using ServiceLayer.CodeCalledInStartup;
 
 namespace ServiceLayer.AuthoriseSetup
 {
