+++ ServiceLayer/AuthoriseSetup/AuthCookieValidateRefreshClaims.cs
@@ -13,7 +13,6 @@ using DataLayer.EfCode;
 using FeatureAuthorize;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.Extensions.DependencyInjection;
-using ServiceLayer.CodeCalledInStartup;
 
 namespace ServiceLayer.AuthoriseSetup
 {
