+++ Test/UnitTests/DataAuthorizeTests/TestHierarchicalFiltering.cs
@@ -3,7 +3,6 @@
 
 using System.Linq;
 using DataLayer.EfCode;
-using ServiceLayer.CodeCalledInStartup;
 using ServiceLayer.SeedDemo.Internal;
 using Test.FakesAndMocks;
 using TestSupport.EfHelpers;
