+++ Test/UnitTests/DataAuthorizeTests/TestShopData.cs
@@ -6,7 +6,6 @@ using DataLayer.EfCode;
 using DataLayer.MultiTenantClasses;
 using Microsoft.EntityFrameworkCore;
 using PermissionParts;
-using ServiceLayer.CodeCalledInStartup;
 using Test.FakesAndMocks;
 using TestSupport.EfHelpers;
 using Xunit;
