+++ Test/UnitTests/SecurityChecks/CheckEntitiesAreSecure.cs
@@ -3,7 +3,6 @@
 
 using System.Linq;
 using DataLayer.EfCode;
-using ServiceLayer.CodeCalledInStartup;
 using Test.EfHelpers;
 using Test.FakesAndMocks;
 using TestSupport.EfHelpers;
