+++ Test/UnitTests/ShopServices/TestShopServices.cs
@@ -8,7 +8,6 @@ using GenericServices.PublicButHidden;
 using GenericServices.Setup;
 using Microsoft.EntityFrameworkCore;
 using PermissionParts;
-using ServiceLayer.CodeCalledInStartup;
 using ServiceLayer.Shop;
 using Test.FakesAndMocks;
 using TestSupport.EfHelpers;
